requirements.txt
python-dotenv>=1.0.0
langchain-mcp-adapters>=0.2.0
langchain-google-genai>=4.0.0
langgraph>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.30.0
streamlit>=1.28.0
requests>=2.28.0
google-auth-oauthlib>=1.0.0
pydantic>=2.0.0
